kernel: CBS: Adding support for Constant Bandwidth Server (CBS)
This commit introduces kernel support for the Constant
Bandwidth Server (CBS), an extension of the Earliest
Deadline First (EDF) scheduler that allows tasks to be
executed virtually isolated from each other, in a way
that if a task executes for longer than expected it doesn’t
interfere on the execution of the others. In other words,
the CBS prevents that a task misbehavior causes other tasks
to miss their own deadlines.

In a nutshell, the CBS is a work-conserving wrapper for
the tasks that automatically recalculates their deadlines
when they exceed their configured timing constraints, thus
acting as a fail-safe for EDF overruns.

Signed-off-by: Alexander pinheiro paschoaletto <1222703@isep.ipp.pt>